Error message; signature policy default
Upstream says don't do this, but I think they're wrong.  The reasoning
is that if the policy file is missing, the user's system is messed up
and they need to be alerted. But then they install a default policy json file that allows everything.

In a scratch container the file doesn't exist so this blows up. I'm
defaulting, and if the user needs a more restrictive policy they can add
the file, just like before.